+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.1] - 2024-05-22
+- FileLogger: Remove constructor with `kotlinx.io.files.Path`
+  Avoids having to add kotlinx-io as a dependency when using the class
+
 ## [1.3.0] - 2024-05-02
 - Provide FileLogger for persistent logging to a file path with rotation and retention policy support.